+++ app/assets/javascript/scrivito_icon_box/editing/editors/button_picker.js
@@ -14,8 +14,8 @@
       clickFunction: function(cmsField) {
         var text = cmsField.data('picker-content');
 
-        cmsField.addClass('selected');
-        cmsField.siblings().removeClass('selected');
+        cmsField.addClass('activated');
+        cmsField.siblings().removeClass('activated');
 
         return cmsField.scrivito('save', text);
       },