+++ cms/migrate/0_create_icon_box_widget.rb
@@ -9,7 +9,7 @@ class CreateIconBoxWidget < ::Scrivito::Migration
         {:name=>"icon", :type=>"string"},
         {:name=>"color", :type=>"string"},
         {:name=>"background_color", :type=>"string"},
-        {:name=>"size", :type=>"enum", :values=>["small","medium","large"]},
+        {:name=>"size", :type=>"enum", :values=>["small","medium","large"]}
       ]
     )
   end